M.A.J (version2)
Hors-Forfait (good)
Frais-Forfait (good)

+++ include/section_consultation.php
@@ -3,7 +3,7 @@
     <h2>Consultation des frais</h2>
     <div class="moisEtat">
       <div class="element">
-        <form class="form-inline" action="index.php" method="post" id="consultationMois">
+        <form class="form-inline" action="index.php?aller=Consultation" method="post" id="consultationMois">
           <div class="form-group">
             <label for="moisConsult"> Consultez votre fiche de frais au mois de : </label>
             <select class="form-control" id="moisConsult" name="moisConsult">
@@ -34,19 +34,14 @@
 
 
 <?php //Mofification de la table LigneFraisForfait
-/*
+
   if (isset($_POST['ModifQuantite']) && isset($_POST['ModifIdFraisForfait'])) {
 
     $ModifQuantitéFraisForfait = $_POST['ModifQuantite'];
     $ModifIdFraisForfait = $_POST['ModifIdFraisForfait'];
     $mois = date('F');
 
-    try {
-      $bdd = getBdd();
-    }
-    catch(Exception $e){
-      die('Erreur : '. $e->getMessage());
-    }
+    $bdd = getBdd();
 
     $ModifQuantité = $bdd -> prepare('UPDATE LigneFraisForfait SET quantite = :quantite WHERE idVisiteur = :id AND idFraisForfait= :idFraisForfait AND mois = :mois');
     $ModifQuantité -> execute(array(':quantite' => $ModifQuantitéFraisForfait, ':id' => $id, ':idFraisForfait' => $ModifIdFraisForfait, ':mois' => $mois));
@@ -59,12 +54,12 @@
     }
 
   }
-*/
+
 ?>
 
 
 <?php //Mofification de la table LigneFraisHorsForfait
-/*
+
   if (isset($_POST['ModifDate']) && isset($_POST['ModifDescription']) && isset($_POST['ModifMontant'])) {
 
     $ModifDate = $_POST['ModifDate'];
@@ -73,12 +68,7 @@
     $ModifIdFraisHorsForfait = $_POST['ModifIdFraisHorsForfait'];
     $mois = date('F');
 
-    try {
-      $bdd = getBdd();
-    }
-    catch(Exception $e){
-      die('Erreur : '. $e->getMessage());
-    }
+    $bdd = getBdd();
 
     $ModifQuantitéFraisHorsForfait = $bdd -> prepare('UPDATE LigneFraisHorsForfait SET libelle = :libelle, dateHF = :dateHF, montant = :montant WHERE idVisiteur = :id AND id= :idFraisHorsForfait AND mois = :mois');
     $ModifQuantitéFraisHorsForfait -> execute(array(':libelle' => $ModifDescription, ':dateHF' => $ModifDate, ':montant' => $ModifMontant, ':id' => $id, ':idFraisHorsForfait' => $ModifIdFraisHorsForfait, ':mois' => $mois));
@@ -91,26 +81,22 @@
     }
 
   }
-*/
+
 ?>
 
 
 <?php //Supression d'une ligne de la table LigneFraisHorsForfait par méthode GET
-/*if (isset($_GET['supp']))
+if (isset($_GET['supp']))
 {
 	if($_GET['supp'] == 'ok')
 	{
-    try {
-      $bdd = getBdd();
-    }
-    catch(Exception $e){
-      die('Erreur : '. $e->getMessage());
-    }
+    
+    $bdd = getBdd();
 
   		$id = $_GET['id']; //Récupération de l'id de la ligne du tableau
   		$supprimer = $bdd -> prepare("DELETE FROM LigneFraisHorsForfait WHERE id = :id AND idVisiteur = :idVisiteur");
   		$supprimer -> execute(array(':id' => $id, ':idVisiteur' => $_SESSION['id']));
   	}
 
-	}*/
+	}
 ?>
